+++ private/bufpkg/bufconfig/module_config.go
@@ -36,6 +36,7 @@ func init() {
 	DefaultModuleConfigV1, err = newModuleConfig(
 		".",
 		nil,
+		nil,
 		map[string][]string{
 			".": {},
 		},
@@ -47,6 +48,7 @@ func init() {
 	}
 	DefaultModuleConfigV2, err = newModuleConfig(
 		".",
+		nil,
 		nil,
 		map[string][]string{
 			".": {},
@@ -78,6 +80,19 @@ type ModuleConfig interface {
 	//
 	// This may be nil.
 	ModuleFullName() bufmodule.ModuleFullName
+	// Includes says the module only includes these paths. The paths are relative to DirPath.
+	// An empty Includes means to include everything.
+	//
+	// This may be set only in v2, always empty for v1beta1 and v1 buf.yamls.
+	//
+	// If both Includes and RootToExcludes are set, i.e. len(Includes) > 0 and
+	// len(RootToExcludes["."]) > 0, the proto files that are considered in the module
+	// are proto files that satisfy BOTH conditions:
+	//  - being inside ONE of the paths in Includes
+	//  - being inside NONE of the paths in RootToExcludes["."]
+	//
+	// Not a copy, do not modify.
+	Includes() []string
 	// RootToExcludes contains a map from root to the excludes for that root.
 	//
 	// Roots are the root directories within a bucket to search for Protobuf files.
@@ -120,6 +135,7 @@ type ModuleConfig interface {
 func NewModuleConfig(
 	dirPath string,
 	moduleFullName bufmodule.ModuleFullName,
+	includes []string,
 	rootToExcludes map[string][]string,
 	lintConfig LintConfig,
 	breakingConfig BreakingConfig,
@@ -127,6 +143,7 @@ func NewModuleConfig(
 	return newModuleConfig(
 		dirPath,
 		moduleFullName,
+		includes,
 		rootToExcludes,
 		lintConfig,
 		breakingConfig,
@@ -138,6 +155,7 @@ func NewModuleConfig(
 type moduleConfig struct {
 	dirPath        string
 	moduleFullName bufmodule.ModuleFullName
+	includes       []string
 	rootToExcludes map[string][]string
 	lintConfig     LintConfig
 	breakingConfig BreakingConfig
@@ -147,6 +165,7 @@ type moduleConfig struct {
 func newModuleConfig(
 	dirPath string,
 	moduleFullName bufmodule.ModuleFullName,
+	includes []string,
 	rootToExcludes map[string][]string,
 	lintConfig LintConfig,
 	breakingConfig BreakingConfig,
@@ -196,6 +215,7 @@ func newModuleConfig(
 	return &moduleConfig{
 		dirPath:        dirPath,
 		moduleFullName: moduleFullName,
+		includes:       includes,
 		rootToExcludes: newRootToExcludes,
 		lintConfig:     lintConfig,
 		breakingConfig: breakingConfig,
@@ -208,6 +228,10 @@ func (m *moduleConfig) DirPath() string {
 
 func (m *moduleConfig) ModuleFullName() bufmodule.ModuleFullName {
 	return m.moduleFullName
+}
+
+func (m *moduleConfig) Includes() []string {
+	return m.includes
 }
 
 func (m *moduleConfig) RootToExcludes() map[string][]string {